Hoist networks to their own file, add aurora
