+++ src/components/NetworkDisplay.tsx
@@ -1,9 +1,9 @@
 import { Button, Typography, useTheme } from "@mui/material"
-import { ChainId, IS_L2_SUPPORTED } from "../constants"
+import { ChainId, NETWORK_LABEL } from "../constants/networks"
 import { UnsupportedChainIdError, useWeb3React } from "@web3-react/core"
 
 import CircleTwoToneIcon from "@mui/icons-material/CircleTwoTone"
-import { NETWORK_LABEL } from "../constants/networks"
+import { IS_L2_SUPPORTED } from "../constants"
 import React from "react"
 
 interface NetworkDisplayProps extends React.ComponentPropsWithoutRef<"button"> {
