+++ src/components/SwapPage.tsx
@@ -9,13 +9,14 @@ import {
   Paper,
   Typography,
 } from "@mui/material"
-import { ChainId, SWAP_TYPES, getIsVirtualSwap } from "../constants"
 import React, { ReactElement, useMemo, useState } from "react"
+import { SWAP_TYPES, getIsVirtualSwap } from "../constants"
 import { formatBNToPercentString, formatBNToString } from "../utils"
 
 import AdvancedOptions from "./AdvancedOptions"
 import { AppState } from "../state/index"
 import { BigNumber } from "@ethersproject/bignumber"
+import { ChainId } from "../constants/networks"
 import { Close } from "@mui/icons-material"
 import ConfirmTransaction from "./ConfirmTransaction"
 import Dialog from "./Dialog"
